+++ receiver/CHANGES.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Improve server to receive POST request
+
 ## 0.0.0
 
 - Initial commit